requirements.txt
boto3==1.15.2
botocore==1.18.2
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-fontawesome-5==1.0.18
django-storages==1.10.1
django-taggit==1.3.0
django-utils-six==2.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.3
Pillow==6.2.0
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.3.0
six==1.12.0
stripe==2.51.0
urllib3==1.25.10
whitenoise==4.1.2
